Move objects into their own classes.
Node class contain spines and leaves. Edge class contain edges between nodes.

The idea is to keep main() as clean as possible and separate code that generates graphs from code that generates configuration.

To do this the Node class contain information about how to draw it via the generate() function.

It also returns the data required by the jinja template to generate the config via the template() function.